move files around and remove uneeded code
