+++ .changeset/smart-years-sparkle.md
@@ -1,0 +1,5 @@
+---
+"@zus-health/ctw-component-library": patch
+---
+
+Simplify condition history component.
