+++ src/components/content/CCDA/render-document-button.tsx
@@ -1,28 +1,27 @@
+import type { BinaryDocument } from "../condition-history/conditions-history";
 import { DocumentButton } from "./document-button";
 import { CollapsibleDataListProps } from "@/components/core/collapsible-data-list";
-import { BinaryDocumentData, SourceDocumentMap } from "@/fhir/conditions";
+import { SourceDocumentMap } from "@/fhir/conditions";
 
 export type RenderDocumentButtonProps = {
   idMap: SourceDocumentMap;
   entry: CollapsibleDataListProps;
-  setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  setRawBinary: React.Dispatch<
-    React.SetStateAction<BinaryDocumentData | undefined>
-  >;
+  updateBinaryDocumentState: React.Dispatch<Partial<BinaryDocument>>;
 };
 
 export const RenderDocumentButton = ({
   idMap,
   entry,
-  setIsModalOpen,
-  setRawBinary,
+  updateBinaryDocumentState,
 }: RenderDocumentButtonProps) => (
   <>
     {idMap.get(entry.id) && idMap.get(entry.id)?.isBinary && (
       <DocumentButton
         onClick={() => {
-          setIsModalOpen(true);
-          setRawBinary(idMap.get(entry.id));
+          updateBinaryDocumentState({
+            isModalOpen: true,
+            rawBinary: idMap.get(entry.id),
+          });
         }}
       />
     )}
